backend/consultant/requirements.txt
fastapi
uvicorn
python-dotenv
pymysql
DBUtils
openai
httpx
openai-agents
pystun3
redis
